change organizer reducers to k-redux-factory
