+++ src/components/form/itemsWithModal/itemsWithModal.container.js
@@ -1,11 +1,10 @@
 import { inject } from 'k-ramel/react'
 
-import { openModal, closeModal } from 'redux/ui/modal'
 import ItemsWithModal from './itemsWithModal'
 
 const mapStore = store => ({
-  openModal: id => () => store.dispatch(openModal(id)),
-  closeModal: () => store.dispatch(closeModal()),
+  openModal: id => () => store.ui.modal.set({ openedModal: id }),
+  closeModal: () => store.ui.modal.set({ openedModal: undefined }),
 })
 
 export default (name, Form) => inject(mapStore)(ItemsWithModal(name, Form))
