+++ src/components/modal/modal.container.js
@@ -1,13 +1,12 @@
 import { inject } from 'k-ramel/react'
 
 import withTheme from 'styles/themes/withTheme'
-import { isModalOpened, closeModal } from 'redux/ui/modal'
 
 import Modal from './modal'
 
 const mapStore = (store, { id }) => ({
-  opened: isModalOpened(store.getState())(id),
-  onClose: () => store.dispatch(closeModal()),
+  opened: store.ui.modal.openedModal === id,
+  onClose: () => store.ui.modal.set({ openedModal: undefined }),
 })
 
 export default inject(mapStore)(withTheme(Modal))
