--- src/components/toaster/index.js
@@ -1,1 +1,0 @@
-export { default } from './toaster.container'
