--- src/components/toaster/toast/index.js
@@ -1,1 +1,0 @@
-export { default } from './toast'
