--- src/components/toaster/toast/toast.css
@@ -1,37 +1,0 @@
-.toast {
-  padding: 1em;
-  margin-bottom: 1em;
-  min-width: 300px;
-  max-width: 600px;
-  box-shadow: 0 10px 20px rgba(0, 0, 0, 0.19), 0 6px 6px rgba(0, 0, 0, 0.23);
-  border-radius: 3px;
-  border: 1px solid var(--grey-color);
-  background: #fff;
-  color: var(--primary-color);
-}
-
-.toast-default {
-  color: var(--primary-color);
-}
-
-.toast-info {
-  color: var(--primary-color);
-}
-
-.toast-success {
-  border: none;
-  color: #fff;
-  background: var(--success-color);
-}
-
-.toast-warning {
-  color: var(--warning-color);
-}
-
-.toast-error {
-  color: var(--error-color);
-}
-
-.toast-question {
-  color: var(--question-color);
-}
