--- src/components/toaster/toaster.css
@@ -1,6 +1,0 @@
-.toaster {
-  position: fixed;
-  bottom: 0;
-  left: 50%;
-  transform: translateX(-50%);
-}
