+++ src/redux/ui/index.js
@@ -1,11 +1,9 @@
 import modal from './modal'
-import toaster from './toaster'
 import speaker from './speaker'
 import organizer from './organizer'
 
 export default {
   modal,
-  toaster,
   speaker,
   organizer,
 }
