--- src/redux/ui/modal/index.js
@@ -1,3 +1,0 @@
-export { default } from './modal'
-export * from './modal'
-export * from './modal.selectors'
