--- src/redux/ui/modal/modal.selectors.js
@@ -1,3 +1,0 @@
-export const getModal = state => state.ui.modal
-
-export const isModalOpened = state => id => getModal(state).openedModal === id
