+++ src/redux/ui/organizer/myEvents.js
@@ -1,0 +1,3 @@
+import { keyValue } from 'k-ramel'
+
+export default keyValue({ key: 'id' })
