--- src/redux/ui/organizer/myEvents/index.js
@@ -1,1 +1,0 @@
-export { default } from './myEvents'
