--- src/redux/ui/organizer/myEvents/myEvents.js
@@ -1,8 +1,0 @@
-import { keyValue } from 'k-ramel'
-
-export default keyValue({
-  key: 'id',
-  name: 'myEvents',
-  path: 'ui.organizer',
-  prefix: 'organizer',
-})
