+++ src/redux/ui/organizer/proposal.selectors.js
