--- src/redux/ui/organizer/proposal/index.js
@@ -1,2 +1,0 @@
-export { default } from './proposal'
-export * from './proposal.selectors'
