+++ src/redux/ui/organizer/proposals.js
@@ -1,0 +1,9 @@
+import { simpleObject } from 'k-ramel'
+
+const defaultData = {
+  categories: '',
+  formats: '',
+  sorting: 'newest',
+}
+
+export default simpleObject({ defaultData })
