--- src/redux/ui/organizer/proposals/filters.js
@@ -1,14 +1,0 @@
-import { simpleObject } from 'k-ramel'
-
-const defaultData = {
-  categories: '',
-  formats: '',
-  sorting: 'newest',
-}
-
-export default simpleObject({
-  path: 'ui.organizer.proposals',
-  name: 'filters',
-  prefix: 'proposals',
-  defaultData,
-})
