--- src/redux/ui/organizer/proposals/index.js
@@ -1,5 +1,0 @@
-import filters from './filters'
-
-export default {
-  filters,
-}
