--- src/redux/ui/toaster/index.js
@@ -1,3 +1,0 @@
-export { default } from './toaster'
-export * from './toaster'
-export * from './toaster.selectors'
