--- src/redux/ui/toaster/toaster.selectors.js
@@ -1,2 +1,0 @@
-/* eslint-disable import/prefer-default-export */
-export const getToasts = state => state.ui.toaster
