+++ src/sagas/index.js
@@ -3,7 +3,6 @@ import { all } from 'redux-saga/effects'
 import firebaseSaga from './firebase'
 import authSaga from './auth'
 import eventsSaga from './events'
-import toasterSaga from './toaster'
 import userSaga from './user'
 import talksSaga from './talks'
 import speakerSaga from './speaker'
@@ -17,7 +16,6 @@ export default function* sagas() {
       firebaseSaga(),
       authSaga(),
       eventsSaga(),
-      toasterSaga(),
       userSaga(),
       talksSaga(),
       speakerSaga(),
