--- src/sagas/toaster/index.js
@@ -1,1 +1,0 @@
-export { default } from './toaster.saga'
