+++ src/screens/organizer/proposals/proposalFilters/proposalFilters.container.js
@@ -5,14 +5,14 @@ import ProposalFilters from './proposalFilters'
 
 const mapStore = (store) => {
   const eventId = getRouterParam('eventId')(store.getState())
-  const filters = store.ui.organizer.proposals.filters.get()
+  const filters = store.ui.organizer.proposals.get()
   const { formats, categories } = store.data.events.get(eventId) || {}
   return {
     formats,
     categories,
     filters,
     onChange: ({ target }) => {
-      store.ui.organizer.proposals.filters.update({ [target.id]: target.value })
+      store.ui.organizer.proposals.update({ [target.id]: target.value })
     },
   }
 }
